+++ dflib-jdbc/src/main/java/com/nhl/dflib/jdbc/connector/tx/Tx.java
@@ -52,5 +52,4 @@ public class Tx {
             throw new RuntimeException(e);
         }
     }
-
 }